chore: replaced applyOptions with direct calls
